Preserve valid vote summaries on malformed refreshes
