+++ components/waves/drop/DropVoteDistribution.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import Button from "@/components/utils/button/Button";
-import type { ApiDropVoteDistribution } from "@/generated/models/ApiDropVoteDistribution";
 import type { ApiDropVoter } from "@/generated/models/ApiDropVoter";
 import { ApiDropType } from "@/generated/models/ApiDropType";
 import { buildTooltipId, TOOLTIP_STYLES } from "@/helpers/tooltip.helpers";
@@ -16,14 +15,8 @@ import { useEffect, useId, useRef, useState } from "react";
 import { createPortal } from "react-dom";
 import { Tooltip } from "react-tooltip";
 import type { TooltipRefProps } from "react-tooltip";
+import { getVoteSides, type VoteSide } from "./dropVoteDistribution.helpers";
 import type { DropVoteSummaryState } from "./useDropVoteSummary";
-
-interface VoteSide {
-  readonly positive: boolean;
-  readonly total: number;
-  readonly remainder: number;
-  readonly voters: readonly ApiDropVoter[];
-}
 
 const POSITIVE_PIECES = [
   "tw-bg-green",
@@ -33,96 +26,6 @@ const POSITIVE_PIECES = [
 const NEGATIVE_PIECES = ["tw-bg-red", "tw-bg-red/80", "tw-bg-red/60"] as const;
 const STRIPED_PIECE =
   "tw-bg-[repeating-linear-gradient(135deg,transparent,transparent_5px,currentColor_5px,currentColor_6px)] tw-text-iron-950/10";
-
-function isVoteEntry(entry: unknown): entry is ApiDropVoter {
-  if (
-    entry === null ||
-    typeof entry !== "object" ||
-    !("vote" in entry) ||
-    !("voter" in entry)
-  ) {
-    return false;
-  }
-  const voter = entry.voter;
-  return (
-    typeof entry.vote === "number" &&
-    Number.isSafeInteger(entry.vote) &&
-    voter !== null &&
-    typeof voter === "object" &&
-    "id" in voter &&
-    typeof voter.id === "string" &&
-    voter.id.trim().length > 0
-  );
-}
-
-function getVoteSide(
-  total: number,
-  voters: unknown,
-  positive: boolean
-): VoteSide | null {
-  const direction = positive ? 1 : -1;
-  if (
-    !Number.isSafeInteger(total) ||
-    total * direction < 0 ||
-    !Array.isArray(voters) ||
-    voters.length > 3
-  ) {
-    return null;
-  }
-  const entries: readonly unknown[] = voters;
-  if (
-    !entries.every(isVoteEntry) ||
-    entries.some((entry) => entry.vote * direction <= 0)
-  ) {
-    return null;
-  }
-  const listedTotal = entries.reduce((sum, entry) => sum + entry.vote, 0);
-  if (
-    !Number.isSafeInteger(listedTotal) ||
-    Math.abs(listedTotal) > Math.abs(total)
-  ) {
-    return null;
-  }
-  return {
-    positive,
-    total,
-    remainder: total - listedTotal,
-    voters: [...entries].sort(
-      (left, right) => Math.abs(right.vote) - Math.abs(left.vote)
-    ),
-  };
-}
-
-function getVoteSides(distribution: ApiDropVoteDistribution | undefined) {
-  if (!distribution) {
-    return null;
-  }
-  const positive = getVoteSide(
-    distribution.positive_total,
-    distribution.positive_votes,
-    true
-  );
-  const negative = getVoteSide(
-    distribution.negative_total,
-    distribution.negative_votes,
-    false
-  );
-  if (!positive || !negative) {
-    return null;
-  }
-  const grossTotal = positive.total - negative.total;
-  const voterIds = [...negative.voters, ...positive.voters].map(
-    (entry) => entry.voter.id
-  );
-  if (
-    !Number.isSafeInteger(grossTotal) ||
-    grossTotal <= 0 ||
-    new Set(voterIds).size !== voterIds.length
-  ) {
-    return null;
-  }
-  return [negative, positive];
-}
 
 function getVoterName({ voter }: ApiDropVoter): string {
   const address =
