+++ components/waves/drop/useDropVoteSummary.ts
@@ -3,13 +3,15 @@
 import { QueryKey } from "@/components/react-query-wrapper/ReactQueryWrapper";
 import type { ApiDrop } from "@/generated/models/ApiDrop";
 import type { ApiDropVoteDistribution } from "@/generated/models/ApiDropVoteDistribution";
+import type { ApiDropVoteSummary } from "@/generated/models/ApiDropVoteSummary";
 import { isWsDropUpdateRefData, WsMessageType } from "@/helpers/Types";
 import { useDebouncedQueryRefetch } from "@/hooks/useDebouncedQueryRefetch";
 import { DROP_DETAIL_STALE_TIME_MS } from "@/services/api/drop-api";
 import { fetchDropVoteSummaryByIdV2 } from "@/services/api/drop-vote-summary-api";
 import { useWebSocketMessage } from "@/services/websocket/useWebSocketMessage";
-import { useQuery } from "@tanstack/react-query";
+import { replaceEqualDeep, useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
+import { getVoteSides } from "./dropVoteDistribution.helpers";
 
 export type DropVoteSummaryState =
   | { readonly status: "disabled" }
@@ -28,6 +30,27 @@ export const DISABLED_DROP_VOTE_SUMMARY_STATE = {
   status: "disabled",
 } as const satisfies DropVoteSummaryState;
 
+function preserveValidVoteSummary(previous: unknown, next: unknown): unknown {
+  if (next === null || typeof next !== "object") {
+    return next;
+  }
+  const nextSummary = next as ApiDropVoteSummary;
+  const previousSummary =
+    previous !== null && typeof previous === "object"
+      ? (previous as ApiDropVoteSummary)
+      : undefined;
+  const nextDistribution = nextSummary.vote_distribution;
+  if (
+    nextDistribution &&
+    !getVoteSides(nextDistribution) &&
+    previousSummary &&
+    getVoteSides(previousSummary.vote_distribution)
+  ) {
+    return previousSummary;
+  }
+  return replaceEqualDeep(previous, nextSummary);
+}
+
 export function useDropVoteSummary({
   dropId,
   waveId,
@@ -37,12 +60,13 @@ export function useDropVoteSummary({
   readonly waveId: string;
   readonly enabled: boolean;
 }): DropVoteSummaryState {
-  const { data, isFetching, refetch } = useQuery({
+  const { data, isFetching, refetch } = useQuery<ApiDropVoteSummary>({
     queryKey: [QueryKey.DROP, { drop_id: dropId, view: "vote-summary" }],
     queryFn: ({ signal }) => fetchDropVoteSummaryByIdV2(dropId, signal),
     enabled,
     staleTime: DROP_DETAIL_STALE_TIME_MS,
     retry: 1,
+    structuralSharing: preserveValidVoteSummary,
   });
   const refetchWhenEnabled = useCallback(
     () => (enabled ? refetch() : Promise.resolve()),
@@ -85,14 +109,15 @@ export function useDropVoteSummary({
   const retry = useCallback(() => {
     void refetch();
   }, [refetch]);
+  const candidate = data?.vote_distribution;
 
   if (!enabled) {
     return DISABLED_DROP_VOTE_SUMMARY_STATE;
   }
-  if (data?.vote_distribution) {
+  if (candidate && getVoteSides(candidate)) {
     return {
       status: "ready",
-      voteDistribution: data.vote_distribution,
+      voteDistribution: candidate,
       retry,
     };
   }
